Moved btcFeeSelectionWidget and fee related strings to separate modules
